Conflicts:
	galileevan.com/db.sqlite3

+++ galileevan.com/posts/views.py
@@ -1,11 +1,19 @@
-from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
-from django.shortcuts import render, get_object_or_404, redirect
-from django.http import HttpResponse, HttpResponseRedirect
+from urllib.parse import quote_plus
+
+
 from django.contrib import messages
+from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
+from django.db.models import Q
+from django.http import HttpResponse, HttpResponseRedirect, Http404
+from django.shortcuts import render, get_object_or_404, redirect
+from django.utils import timezone
+
 
 
 from posts.models import Post
 from .forms import PostForm
+
+
 
 # Create your views here.
 # def posts_home(request):
@@ -19,9 +27,15 @@ from .forms import PostForm
 
 
 def posts_create(request):
+	# if not request.user.is_staff or not request.user.is_superuser:
+	# 	raise Http404
+	# if not request.user.is_authenticated():
+	# 	raise Http404
+
 	form = PostForm(request.POST or None, request.FILES or None)
 	if form.is_valid():
 		instance = form.save(commit=False)
+		instance.user = request.user
 		instance.save()
 		messages.success(request, "Successfully Created!")
 		return HttpResponseRedirect(instance.get_absolute_url())
@@ -34,18 +48,35 @@ def posts_create(request):
 	}
 	return render(request, "post_form.html", context)
 
-def posts_detail(request, id):
-	instance = get_object_or_404(Post, id=id)
+def posts_detail(request, slug):
+	instance = get_object_or_404(Post, slug=slug)
+	share_string = quote_plus(instance.content)
 	context = {
 		"title": instance.title,
 		"instance": instance,
+		"share_string": share_string,
 	}
 	return render(request, "post_detail.html", context)
 
 def posts_list(request):
-	queryset_list = Post.objects.all().order_by("-timestamp")
+	today = timezone.now().date()
+
+	if request.user.is_staff or request.user.is_superuser:
+		queryset_list = Post.objects.all().order_by("-timestamp")
+	else:
+		queryset_list = Post.objects.filter(user=request.user)
+	
+	query = request.GET.get("q")
+	if query:
+		queryset_list = queryset_list.filter(
+			Q(title__icontains=query) |
+			Q(content__icontains=query) |
+			Q(user__first_name__icontains=query) |
+			Q(user__last_name__icontains=query)
+		).distinct()
+
 	paginator = Paginator(queryset_list, 5) # Show 25 contacts per page
-	page_request_var = "abc"
+	page_request_var = "page"
 	page = request.GET.get(page_request_var)
 	try:
 	    queryset = paginator.page(page)
@@ -78,8 +109,10 @@ def posts_list(request):
 
 
 
-def posts_update(request, id=None):
-	instance = get_object_or_404(Post, id=id)
+def posts_update(request, slug=None):
+	if not request.user.is_staff or not request.user.is_superuser:
+		raise Http404
+	instance = get_object_or_404(Post, slug=slug)
 	form = PostForm(request.POST or None, request.FILES or None, instance=instance)
 	if form.is_valid():
 		instance = form.save(commit=False)
@@ -96,8 +129,10 @@ def posts_update(request, id=None):
 	}
 	return render(request, "post_form.html", context)
 
-def posts_delete(request, id=None):
-	instance = get_object_or_404(Post, id=id)
+def posts_delete(request, slug=None):
+	if not request.user.is_staff or not request.user.is_superuser:
+		raise Http404
+	instance = get_object_or_404(Post, slug=slug)
 	instance.delete()
 	messages.success(request, "Successfully Deleted!")
 	return redirect("posts:list")
